Added some fixes
